chore(ci): build fix attempt

+++ src/da4ml/_binary/dais/DAISInterpreter.hh
@@ -54,7 +54,7 @@ namespace dais {
 
     class DAISInterpreter {
       private:
-        int32_t n_in, n_out, n_ops, n_tables;
+        size_t n_in, n_out, n_ops, n_tables;
         int32_t max_ops_width = 0, max_inp_width = 0, max_out_width = 0;
         std::vector<int32_t> inp_shifts;
         std::vector<int32_t> out_idxs;
